+++ lib/petitest/test_group.rb
@@ -1,7 +1,5 @@
 module Petitest
   class TestGroup
-    include ::Petitest::Assertions
-
     TEST_METHOD_NAME_PREFIX = "test_"
 
     class << self
@@ -45,6 +43,48 @@ module Petitest
             path: unbound_method.source_location[0],
           )
         end
+      end
+    end
+
+    # @param actual_or_message [Object]
+    # @param message [String, nil]
+    def assert(actual_or_message = nil, message = nil, &block)
+      if block
+        check(
+          message: message,
+          template: "Given block returned falsy",
+          &block
+        )
+      else
+        actual = actual_or_message
+        check(
+          message: message,
+          template: "%{actual} is not truthy",
+          template_variables: {
+            actual: actual,
+          },
+        ) do
+          actual
+        end
+      end
+    end
+
+    private
+
+    # @param message [String, nil]
+    # @param template [String]
+    # @param template_variables [Hash, nil]
+    def check(
+      message:,
+      template:,
+      template_variables: {}
+    )
+      unless yield
+        raise ::Petitest::AssertionFailureError.new(
+          additional_message: message,
+          template: template,
+          template_variables: template_variables,
+        )
       end
     end
   end